Move unit-tests compatible test to unit/ folder
