+++ test/unit/messages/CampaignSignupMessage.test.js
