+++ test/unit/messages/CampaignSignupPostMessage.test.js
