+++ test/unit/messages/UserMessage.test.js
