+++ test/unit/workers/GambitCampaignSignupRelayWorker.test.js
